refact: Page create dialog as dialog controller
